missing cast

+++ assembler/src/handle.c
@@ -477,7 +477,7 @@ bool handle_branch_cond(const struct TokenList *tokenList,
 		offset = (symbolTable->symbols[symbol_num].address -
 				 current_address - 2) >>
 			1;
-		if (offset > 511 || (int16_t)offset < -512) {
+		if ((int16_t)offset > 511 || (int16_t)offset < -512) {
 			printf("Warning: %d: Branch target too far away from %s.\n",
 				next3->line, next3->str);
 		}
